pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "veo-4-api"
version = "0.1.0"
authors = [
  { name="Anil Matcha" },
]
description = "A comprehensive Python wrapper for the Veo 4 API by Google DeepMind."
readme = "README.md"
requires-python = ">=3.7"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "requests",
    "python-dotenv",
    "mcp[cli]"
]

[tool.setuptools]
py-modules = ["veo4_api", "mcp_server"]
